Seperated Include Full Checkbox into its own component
Ryan brought up the point about the include full checkbox not appearing if there were no honors/async/remote sections so I made it so it appeared regardless. However copy pasting it into the second return statement made the code pretty messy so I figured I might as well seperate it into its own component becuase it looked cleaner and that way if anyone needed a checkbox in the future for some reason they could also make one pretty easily.
