+++ autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/BasicSelect/BasicSelect.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import {
-  Typography, FormLabel, Checkbox,
+  Typography, FormLabel,
 } from '@material-ui/core';
-import { useSelector, useDispatch } from 'react-redux';
-import { updateCourseCard } from '../../../../../../redux/actions/courseCards';
+import { useSelector } from 'react-redux';
 import { RootState } from '../../../../../../redux/reducer';
 import * as styles from './BasicSelect.css';
 import BasicOptionRow from './BasicOptionRow';
+import BasicCheckbox from './BasicCheckbox';
 
 interface BasicSelectProps {
   id: number;
@@ -23,10 +23,6 @@ const BasicSelect: React.FC<BasicSelectProps> = ({ id }) => {
   const hasAsynchronous = useSelector<RootState, boolean>(
     (state) => state.courseCards[id].hasAsynchronous || false,
   );
-  const includeFull = useSelector<RootState, boolean>(
-    (state) => state.courseCards[id].includeFull || false,
-  );
-  const dispatch = useDispatch();
 
   // shows placeholder text if no course is selected
   if (!course) {
@@ -40,9 +36,17 @@ const BasicSelect: React.FC<BasicSelectProps> = ({ id }) => {
   // show placeholder message if there are no special sections to filter
   if (!hasHonors && !hasRemote && !hasAsynchronous) {
     return (
-      <Typography className={styles.placeholderText} color="textSecondary">
+      <>
+        <FormLabel>Options</FormLabel>
+        <table className={styles.tableContainer}>
+          <tbody>
+            <BasicCheckbox id={id} value="includeFull" label="Include Full Sections" />
+          </tbody>
+        </table>
+        <Typography className={styles.placeholderText} color="textSecondary">
         There are no honors or remote sections for this class
-      </Typography>
+        </Typography>
+      </>
     );
   }
 
@@ -51,23 +55,7 @@ const BasicSelect: React.FC<BasicSelectProps> = ({ id }) => {
       <FormLabel>Options</FormLabel>
       <table className={styles.tableContainer}>
         <tbody>
-          <tr>
-            <td>
-              <Typography variant="body1" style={{ paddingRight: 8 }}>
-            Include Full Sections:
-              </Typography>
-            </td>
-            <td>
-              <Checkbox
-                color="primary"
-                style={{ padding: 0 }}
-                checked={includeFull}
-                onChange={(): void => {
-                  dispatch(updateCourseCard(id, { includeFull: !includeFull }));
-                }}
-              />
-            </td>
-          </tr>
+          <BasicCheckbox id={id} value="includeFull" label="Include Full Sections" />
           {hasHonors
             ? <BasicOptionRow id={id} value="honors" label="Honors" />
             : null}
